pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ringmaster"
version = "0.1.0"
description = "GPU workstation AI task orchestrator for home networks"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "click>=8.1",
    "dbus-next>=0.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
ringmaster = "ringmaster.cli.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
